% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_var_groups.R
\name{calc_var_groups}
\alias{calc_var_groups}
\title{Calculate variance in defined groups}
\usage{
calc_var_groups(exp_matrix, groups)
}
\arguments{
\item{exp_matrix}{An normalized expression matrix.}

\item{groups}{A factor with the experiment groups in the same order as it appears on the expression matrix.}
}
\value{
A data.frame with group variances
}
\description{
Calculate variance in defined groups
}
