% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_var_total.R
\name{calc_var_total}
\alias{calc_var_total}
\title{Calculate the total variance}
\usage{
calc_var_total(exp_matrix)
}
\arguments{
\item{exp_matrix}{An normalized expression matrix.}
}
\value{
A numeric vector with the total variance for each probe.
}
\description{
Calculate the total variance
}
