% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_var.R
\name{compare_variances}
\alias{compare_variances}
\title{Compare variances}
\usage{
compare_variances(var_groups, var_total)
}
\arguments{
\item{var_groups}{A dataframe with the variances of each group}

\item{var_total}{A numeric vector with the total variance}
}
\value{
A named vector with the proportion of probes from groups with greater variance
than the total variance.
}
\description{
Compare the variance of each group with the total variance
}
