% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create report}
\usage{
create_report(celfiles = NULL, pheno_data = NULL, sep = NULL,
  filenames = NULL, components = NULL, group = NULL, batch = NULL,
  type = NULL, ...)
}
\arguments{
\item{celfiles}{Path to the celfiles directory}

\item{pheno_data}{File name with metadata information from the experiment}

\item{sep}{\code{pheno_data} file separated format}

\item{filenames}{A vector containg the names of .CEL files, if only a subset of files are needed to be imported.}

\item{components}{A vector of length 2 containing which principal components to plot}

\item{batch}{Column name corresponding to each replicates on pheno data provided to the \code{import_celfiles()} function (optional)}

\item{type}{Column name corresponding to the types of samples (control and treated samples) on pheno data provided to the \code{import_celfiles()} function (optional)}

\item{...}{Other arguments to be passed to \code{knitr::render()} function. Useful arguments:
\code{output_file}, \code{output_dir}, etc. See \code{?render}}

\item{groups}{Column name corresponding to the groups of samples annotation on pheno data provided to the \code{import_celfiles()} function}
}
\value{
Produces an html file with quality assessment plots.
}
\description{
Create quality assessment report of Affymetrix CEL files
}
\details{
\code{pheno_data} is a table containing basic information about the experiment, such as GSMid and sample groups.
It must have at least these two columns (GSMid and sample groups). For PCA plot, one could provide the type of each samples (control or treated)
and the batches (replicates) of each sample.
}
