% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tg_dict.R
\name{create_tg_dict}
\alias{create_tg_dict}
\title{Create TG dictionary}
\usage{
create_tg_dict(ensembleGeneId, dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{ensembleGeneId}{list of ensemble GeneId that have counts.}

\item{dataset}{ENSEMBL dataset name. A dataset name can be retrieved with list_datasets() function.}
}
\description{
Create a dictionary to transcriptogramer package
}
\details{

}