% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_celfiles.R
\name{import_celfiles}
\alias{import_celfiles}
\title{Import celfiles (Affymetrix)}
\usage{
import_celfiles(celfiles_path, pheno_data, sep, filenames = NULL)
}
\arguments{
\item{celfiles_path}{Path to the celfiles directory}

\item{pheno_data}{File name with metadata information from the experiment}

\item{sep}{\code{pheno_data} separated format}

\item{filenames}{a vector containg the names of .CEL files, if only a subset of files are needed to be imported. (optional)}
}
\value{
An AffyBatch object with imported celfiles
}
\description{
Import celfiles (Affymetrix)
}
\details{
\code{pheno_data} is a table containing basic information about the experiment, such as GSMid and sample groups.
It must have at least these two columns (GSMid and sample groups). For PCA plot, one could provide the type of each samples (control or treated)
and the batches (replicates) of each sample.
}
