% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_norm.R
\name{import_norm}
\alias{import_norm}
\title{Import and normalize}
\usage{
import_norm(celfiles_path, pheno_data, sep, filenames = NULL, ...)
}
\arguments{
\item{celfiles_path}{Path to the celfiles directory}

\item{pheno_data}{File name with metadata information from the experiment}

\item{sep}{\code{pheno_data} separated format}

\item{filenames}{a vector containg the names of .CEL files, if only a subset of files are needed to be imported. (optional)}

\item{...}{Other parameters to be passed to \code{rma()} function. See \code{?rma}}
}
\value{
An ExpressionSet object with normalized expression values
}
\description{
Import CEL files (Affymetrix) and normalize with RMA method
}
