% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_attributes.R
\name{list_all_attributes}
\alias{list_all_attributes}
\title{List all attributes}
\usage{
list_all_attributes(dataset, db = "main")
}
\arguments{
\item{dataset}{ENSEMBL dataset name. A dataset name can be retrieved with list_datasets() function.}

\item{db}{One of the following ENSEMBL databases: "main" for ENSEMBL_MART_ENSEMBL database (the main database in ENSEMBL, provides information for the majority of species)
and "mouse" for ENSEMBL_MART_MOUSE database (which provides specific information of mouse strains)}
}
\value{
Dataframe containing all attributes from a dataset.
}
\description{
List all attributes
}
