% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_celfiles.R
\name{list_celfiles}
\alias{list_celfiles}
\title{List celfiles in a given directory}
\usage{
list_celfiles(path)
}
\arguments{
\item{path}{Directory path to look for .CEL files}
}
\value{
A character vector correponding to cel file names in a given directory.
}
\description{
List celfiles in a given directory
}
