% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List ENSEMBL datasets}
\usage{
list_datasets(db = "main")
}
\arguments{
\item{db}{One of the following ENSEMBL databases: "main" for ENSEMBL_MART_ENSEMBL database (the main database in ENSEMBL, provides information for the majority of species)
and "mouse" for ENSEMBL_MART_MOUSE database (which provides specific information of mouse strains)}
}
\description{
List ENSEMBL datasets
}
