% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_platforms.R
\name{list_platforms}
\alias{list_platforms}
\title{List microarray platforms}
\usage{
list_platforms(dataset, db = "main")
}
\arguments{
\item{dataset}{ENSEMBL dataset name. A dataset name can be retrieved with list_datasets() function.}

\item{db}{One of the following ENSEMBL databases: "main" for ENSEMBL_MART_ENSEMBL database (the main database in ENSEMBL, provides information for the majority of species)
and "mouse" for ENSEMBL_MART_MOUSE database (which provides specific information of mouse strains)}
}
\description{
List microarray platforms annotation available for organisms
}
