% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_prop.R
\name{pca_prop}
\alias{pca_prop}
\title{Plot PCs proportions}
\usage{
pca_prop(eset, filename = NULL, ...)
}
\arguments{
\item{eset}{ExpressionSet object with normalized expression values. Commonly the result of RMA normalization}

\item{filename}{File name to save plot (pdf format)}

\item{...}{Other arguments to be passed to pdf() function}
}
\value{
PCs proportions plot
}
\description{
Plot PCs proportions
}
