% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Samples dendrogram}
\usage{
plot_dendrogram(eset, groups, method = "euclidean", filename = NULL,
  ...)
}
\arguments{
\item{eset}{ExpressionSet object with normalized expression values. Commonly the result of RMA normalization}

\item{groups}{Column name corresponding to the groups of samples annotation on pheno data provided to the `import_celfiles` function}

\item{method}{Method to calculate distance between samples. See ?dist()}

\item{filename}{File name to save plot (pdf format)}

\item{...}{Other arguments to be passed to pdf() function}
}
\description{
Samples dendrogram
}
