% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hk.R
\name{plot_hk}
\alias{plot_hk}
\title{Plot housekeeping expression levels}
\usage{
plot_hk(eset, id, actin, gapdh, filename = NULL, ...)
}
\arguments{
\item{eset}{An ExpressionSet object}

\item{id}{Column name for the gene identifier used in eset object}

\item{actin}{Acting gene id}

\item{gapdh}{GAPDH gene id}
}
\description{
Plot housekeeping expression levels
}
