% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA plot}
\usage{
plot_pca(eset, components, groups, type = NULL, batch = NULL,
  filename = NULL, ...)
}
\arguments{
\item{eset}{ExpressionSet object with normalized expression values. Commonly the result of RMA normalization}

\item{components}{A vector of length 2 containing which principal components to plot}

\item{groups}{Column name corresponding to the groups of samples annotation on pheno data provided to the \code{import_celfiles()} function}

\item{type}{Column name corresponding to the types of samples (control and treated samples) on pheno data provided to the \code{import_celfiles()} function (optional)}

\item{batch}{Column name corresponding to each replicates on pheno data provided to the \code{import_celfiles()} function (optional)}

\item{filename}{File name to save plot (pdf format)}

\item{...}{Other arguments to be passed to \code{pdf()} function}
}
\value{
Plot of chosen PCs
}
\description{
PCA plot
}
