% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_probes.R
\name{remove_probes}
\alias{remove_probes}
\title{Remove probes}
\usage{
remove_probes(fdata, probe_col, ref_col = "hgnc_symbol")
}
\arguments{
\item{fdata}{Feature data (dataframe) with unique rows. The output of get_annotation() function}

\item{probe_col}{Probe column name in fdata dataframe}

\item{ref_col}{Column name of reference id to be counted for each probe.}
}
\value{
ExpressionSet object with filtered assayData slot for promiscous probes
}
\description{
Remove promiscuous probes that map to different reference ids (usually gene symbols).
}
