% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_fdata.R
\name{set_fdata}
\alias{set_fdata}
\title{Set feature data}
\usage{
set_fdata(eset, fdata, probe_col, rm_probes = FALSE,
  ref_col = "hgnc_symbol")
}
\arguments{
\item{eset}{ExpressionSet object. Usually the result of the normalization process}

\item{fdata}{A dataframe. Feature data containing at least two columns: probe id and other id (gene symbol, for instance).}

\item{probe_col}{Character vector of length 1. Column name for probes in \code{fdata}}

\item{rm_probes}{Logical. Whether to remove promiscuous probes form feature data dataframe.}

\item{ref_col}{Character vector of length 1. Column name of feature data with the reference id to look for promiscuous probes.}
}
\value{
ExpressionSet object with filtered assayData slot.
}
\description{
Set feature data into ExpressionSet object. Additionaly, it is possible to remove promiscuous probes that
map to different ids.
}
\examples{


}
